Use keyboard controller for hard reset instead of CF9
For some still unknown reason writes to location 0xcf9 do not cause
the Link to reboot, they cause it to shut down instead. While this
will have to be investigated and fixed, this change modifies the code
to have the kernel use the keyboard controller (implemented by the EC
on Link) to restart the system.

Once the 0xcf9 problem is resolved, this change could be reverted.

BUG=chrome-os-partner:8397
TEST=manual
      . typing 'reboot' at the shell causes the system to restart. It
        was shutting down before this change.

Change-Id: I515c87c8ffb57c444bfc3e7074d584e7cbefa87f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18333
Reviewed-by: Randall Spangler <rspangler@chromium.org>