scripts: Show an error if a symbol file is too big
If the symbol file to be uploaded to the symbol server is close to the limit
at which the symbol server will reject it, show an error so we can do
something about it before it's too late.  This also prints the special
"@@@STEP_WARNINGS@@@" line to make the buildbot show the step as orange.

BUG=chromium-os:19194
TEST=Manually ran upload_symbols to make sure it worked with/without an error

Change-Id: I7942ba20f7bc83d66036f9f9fe66403083b1a1f1
Reviewed-on: https://gerrit.chromium.org/gerrit/20664
Commit-Ready: Michael Krebs <mkrebs@chromium.org>
Tested-by: Michael Krebs <mkrebs@chromium.org>
Reviewed-by: David James <davidjames@chromium.org>