Set a sane signal/noise ratio for cros_generate_breakpad_symbols.
Of the 11713 lines output via this for a mario build, 96% of it
is stating "Using dump_syms.32 for 32-bit file <the-path>".

At one point that may have been useful; now it just obscures errors,
thus only output that info when verbose is turned on.

BUG=None
TEST=./cros_generate_breakpad_symbols; # enjoy the 438 lines of
     # output rather than the 11,700 lines of output.

Change-Id: Iba9d1af3421c6b377af8388446521d106399ce25
Reviewed-on: https://gerrit.chromium.org/gerrit/21925
Tested-by: Brian Harring <ferringb@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Brian Harring <ferringb@chromium.org>