Reduce kernel loglevel to 0
Change the kernel loglevel from 1 to 0 to avoid showing console
messages on the screen during shutdown.

An example is the "Power down." message that is shown when shutting
down. This appears on the screen because the kernel prints the
message with KERN_EMERG (level 0). As such, 0 < 1, and the message
appears on the screen.

BUG=chromium-os:28602
TEST=Tested on lumpy, saw no messages when shutting down.

Change-Id: Id3842c2203f6cc4bf3bc9165d8537f440fffba61
Reviewed-on: https://gerrit.chromium.org/gerrit/22104
Reviewed-by: Olof Johansson <olofj@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Commit-Ready: Sean Paul <seanpaul@chromium.org>