Work around tar/pbzip2 idiocy limiting it to single core.
For the builders/goobuntu, they're running pbzip2 ~1.0.5 w/ tar 1.22;
for whatever reason, that configuration reproducibly limits to single
core for:

tar -I /usr/bin/pbzip2 -xf /the/sdk

This is annoying; 2 minutes instead of 10s for 48 core builder for
example.  Thus does *not* occur w/in the chroot (differing versions),
nor for tar=1.26 w/ pbzip2 1.1.6.  The changelogs for both programs
are a bit spartan, but I'm suspecting tar just wasn't feeding it
particularly well (pbzip2 1.0.5 will parallelize if stdin is a pipe).

Regardless, we either try to force everyone to upgrade, or we just
use a form that behaves fine, which is what this CL does.

BUG=chromium-os:31320
TEST=manual validation of it.

Change-Id: I77a434bd2c70873459cbf373192fe73feadb2547
Reviewed-on: https://gerrit.chromium.org/gerrit/23811
Tested-by: Brian Harring <ferringb@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Brian Harring <ferringb@chromium.org>