build_kernel_image: arm: Remove vmalloc command line argument
We were putting 'vmalloc=234MB' in the command line of all ARM boards
even though that's really board-specific.  This shell script really
can't do board-specific stuff, just architecture-specific stuff (and
it would be really hard to change that).  ...so we're just going to
remove the vmalloc part and we'll have to add it back in in some other
way.

Note that the vmalloc was only there for tegra2 devices and there is
no active development going on there.  See http://crosbug.com/33133
for the task to fix the regression and re-add vmalloc somehow for
tegra boards.

BUG=chromium-os:24808
TEST=Built an image and didn't see vmalloc

Change-Id: I6d45143dec5e7ae1bdd207241d275be261085f90
Reviewed-on: https://gerrit.chromium.org/gerrit/28875
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>