scripts: Dump symbols for executables that are not user readable
If an executable was only readable by root, we would skip it with a "Binary
is not user readable" message because the 'file' command was not running as
root.  This kept us from generating Breakpad symbols for Xorg, which is now
a top crasher for some boards.

BUG=chrome-os-partner:11791
TEST=Manually ran the script on some files

Change-Id: Id8a13c2968223e54cb1f6c5c3211e9bc2ca89057
Reviewed-on: https://gerrit.chromium.org/gerrit/29839
Tested-by: Michael Krebs <mkrebs@chromium.org>
Reviewed-by: Brian Harring <ferringb@chromium.org>
Commit-Ready: Michael Krebs <mkrebs@chromium.org>