Remove external_extensions.json from test images
Remove external_extensions.json to further reduce the differences between
test and base images.

/opt/google/chrome/extensions/external_extensions.json was a symlink to
/usr/local/autotest/tests/external_extensions.json (broken by default)
so that a test could put an extension at the destination to make chrome
load it at login. This was an ugly hack from the time when we didn't have
a way to install extensions during tests on chromeos.

power_LoadTest has been updated to not need this hack.

BUG=None
TEST=None

Change-Id: Ibee30315aef725d69725fcf9c5ea72687d030f4e
Reviewed-on: https://gerrit.chromium.org/gerrit/33255
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Nirnimesh <nirnimesh@chromium.org>
Tested-by: Nirnimesh <nirnimesh@chromium.org>