feat(image_to_vm): Initial VirtualBox and vagrant support.
VirtualBox should work but is untested at the moment. Need to setup a
new system that has a kernel it will cooperate with.

vagrant is started but fails because it doesn't tolerate running when
VirtualBox is complaining about no kernel modules being present as it
does on my system. We'll need to figure out a solution here because the
CoreOS-based build hosts won't have VirtualBox support either.